// src/app/core/api/v1/model/meetingTodoAggregate.ts
/**
 * Quarkus CIS Home service
 * Quarkus ChristInformationSystem
 *
 * The version of the OpenAPI document: 1.0
 * Contact: dominik@christ-rlp.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeetingTodo } from './meetingTodo';
import { MeetingId } from './meetingId';


export interface MeetingTodoAggregate { 
    meetingId?: MeetingId;
    todos?: Array<MeetingTodo>;
    todoCounter?: number;
    finishedTodoCounter?: number;
    overdueCounter?: number;
}